feat : [점진적 리팩터링] 수정된 플레이어, 딜러 객체로 교체 완료
