+++ src/main/kotlin/blackjack/domain/participant/Participant.kt
@@ -1,17 +1,28 @@
 package blackjack.domain.participant
 
+import blackjack.domain.card.PlayerCards
 import blackjack.domain.card.TrumpCard
 
-interface Participant {
-    val cards: List<TrumpCard>
+abstract class Participant {
+    private var _cards = PlayerCards(emptySet())
+    val cards get() = _cards
 
-    val sumOfCards: Int
+    fun cardSize(): Int = _cards.size()
 
-    fun addCard(card: TrumpCard)
+    fun addCard(card: TrumpCard) {
+        _cards = _cards.add(card)
+    }
 
-    fun hasAce(): Boolean
+    fun hasAce(): Boolean = _cards.hasAce()
 
-    fun isBust(): Boolean
+    protected fun sumOfCards(): Int = cards.sumOfCards()
 
-    fun sum(): Int
+    abstract fun isBust(): Boolean
+
+    abstract fun totalScore(): Int
+
+    companion object {
+        const val DEALER_MAX_SCORE = 16
+        const val ACE_EXTRACT_SCORE = 10
+    }
 }
