+++ src/test/kotlin/blackjack/domain/DealerTest.kt
@@ -4,8 +4,10 @@ import blackjack.domain.card.Shape
 import blackjack.domain.card.Tier
 import blackjack.domain.card.TrumpCard
 import blackjack.domain.participant.Dealer
+import blackjack.fixture.bustTrumpCardFixture
 import blackjack.fixture.trumpCardFixture
 import org.assertj.core.api.Assertions.assertThat
+import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 
@@ -23,24 +25,50 @@ class DealerTest {
         fixture.forEach {
             dealer.addCard(it)
         }
-        assertThat(dealer.cards).containsExactly(*fixture.toTypedArray())
+        assertThat(dealer.cards.items).containsExactly(*fixture.toTypedArray())
+    }
+
+    @Test
+    fun `딜러가 에이스 카드가 없으면 에이스 카드가 없음을 반환한다`() {
+        val expected = dealer.hasAce()
+
+        assertEquals(expected, false)
+    }
+
+    @Test
+    fun `딜러가 에이스 카드가 있으면 에이스 카드가 있음을 반환한다`() {
+        dealer.addCard(TrumpCard(Tier.ACE, Shape.DIA))
+
+        val expected = dealer.hasAce()
+
+        assertEquals(expected, true)
+    }
+
+    @Test
+    fun `딜러 카드의 카드를 추가하면 사이즈가 증가한다`() {
+        dealer.addCard(TrumpCard(Tier.KING, Shape.DIA))
+        dealer.addCard(TrumpCard(Tier.TEN, Shape.DIA))
+        dealer.addCard(TrumpCard(Tier.TEN, Shape.HEART))
+
+        assertEquals(dealer.cardSize(), 3)
     }
 
     @Test
     fun `딜러 카드의 총합이 21을 초과하면 버스트 된다`() {
-        repeat(3) {
-            dealer.addCard(TrumpCard(Tier.KING, Shape.DIA))
+        bustTrumpCardFixture().forEach {
+            dealer.addCard(it)
         }
-        assertThat(dealer.isBust()).isEqualTo(true)
+
+        assertEquals(dealer.isBust(), true)
     }
 
     @Test
     fun `에이스가 없을 때 총합을 구해서 16을 초과하면 최대 점수를 초과 했음을 반환한다`() {
-        dealer.addCard(TrumpCard(Tier.KING, Shape.DIA))
-        dealer.addCard(TrumpCard(Tier.KING, Shape.HEART))
+        bustTrumpCardFixture().forEach {
+            dealer.addCard(it)
+        }
 
-        val expected = true
-        assertThat(dealer.isOverMaxScore()).isEqualTo(expected)
+        assertEquals(dealer.isOverMaxScore(), true)
     }
 
     @Test
@@ -48,8 +76,7 @@ class DealerTest {
         dealer.addCard(TrumpCard(Tier.ACE, Shape.DIA))
         dealer.addCard(TrumpCard(Tier.KING, Shape.DIA))
 
-        val expected = true
-        assertThat(dealer.isOverMaxScore()).isEqualTo(expected)
+        assertEquals(dealer.isOverMaxScore(), true)
     }
 
     @Test
@@ -57,8 +84,7 @@ class DealerTest {
         dealer.addCard(TrumpCard(Tier.ACE, Shape.DIA))
         dealer.addCard(TrumpCard(Tier.TWO, Shape.DIA))
 
-        val expected = false
-        assertThat(dealer.isOverMaxScore()).isEqualTo(expected)
+        assertEquals(dealer.isOverMaxScore(), false)
     }
 
     @Test
@@ -66,8 +92,7 @@ class DealerTest {
         dealer.addCard(TrumpCard(Tier.ACE, Shape.DIA))
         dealer.addCard(TrumpCard(Tier.KING, Shape.DIA))
 
-        val expected = true
-        assertThat(dealer.isOverMaxScore()).isEqualTo(expected)
+        assertEquals(dealer.isOverMaxScore(), true)
     }
 
     @Test
@@ -75,8 +100,7 @@ class DealerTest {
         dealer.addCard(TrumpCard(Tier.ACE, Shape.DIA))
         dealer.addCard(TrumpCard(Tier.ACE, Shape.HEART))
 
-        val expected = false
-        assertThat(dealer.isOverMaxScore()).isEqualTo(expected)
+        assertEquals(dealer.isOverMaxScore(), false)
     }
 
     @Test
@@ -85,8 +109,7 @@ class DealerTest {
         dealer.addCard(TrumpCard(Tier.ACE, Shape.HEART))
         dealer.addCard(TrumpCard(Tier.NINE, Shape.HEART))
 
-        val expected = true
-        assertThat(dealer.isOverMaxScore()).isEqualTo(expected)
+        assertEquals(dealer.isOverMaxScore(), true)
     }
 
     @Test
@@ -94,15 +117,15 @@ class DealerTest {
         dealer.addCard(TrumpCard(Tier.ACE, Shape.DIA))
         dealer.addCard(TrumpCard(Tier.NINE, Shape.HEART))
 
-        assertThat(dealer.sum()).isEqualTo(20)
+        assertEquals(dealer.totalScore(), 20)
     }
 
     @Test
     fun `에이스 카드를 가지고 버스트 되었으면 카드 총합을 유지한다`() {
-        dealer.addCard(TrumpCard(Tier.ACE, Shape.DIA))
-        dealer.addCard(TrumpCard(Tier.SEVEN, Shape.HEART))
-        dealer.addCard(TrumpCard(Tier.NINE, Shape.HEART))
+        bustTrumpCardFixture().forEach {
+            dealer.addCard(it)
+        }
 
-        assertThat(dealer.sum()).isEqualTo(17)
+        assertEquals(dealer.totalScore(), 30)
     }
 }
